+++ tests/v3/integration.test.js
@@ -14,7 +14,7 @@ let consoleWarnMock;
 
 describe('info.json', () => {
   it('produces a valid info.json', async () => {
-    subject = new Processor(3, `${base}/info.json`, streamResolver, { pathPrefix: 'iiif/3/ab/cd/ef/gh' });
+    subject = new Processor(`${base}/info.json`, streamResolver, { pathPrefix: 'iiif/3/ab/cd/ef/gh' });
     const result = await subject.execute();
     const info = JSON.parse(result.body);
     assert.strictEqual(info.id, 'https://example.org/iiif/3/ab/cd/ef/gh/i');
@@ -24,7 +24,7 @@ describe('info.json', () => {
   });
 
   it('respects max size options', async () => {
-    subject = new Processor(3, `${base}/info.json`, streamResolver, { pathPrefix: 'iiif/3/ab/cd/ef/gh', max: { width: 600 } });
+    subject = new Processor(`${base}/info.json`, streamResolver, { pathPrefix: 'iiif/3/ab/cd/ef/gh', max: { width: 600 } });
     const result = await subject.execute();
     const info = JSON.parse(result.body);
     assert.strictEqual(info.maxWidth, 600);
@@ -36,7 +36,7 @@ describe('info.json', () => {
 describe('quality', () => {
   qualities.forEach((value) => {
     it(`should produce an image with quality ${value}`, async () => {
-      subject = new Processor(3, `${base}/full/max/0/${value}.png`, streamResolver);
+      subject = new Processor(`${base}/full/max/0/${value}.png`, streamResolver);
       const result = await subject.execute();
       assert.strictEqual(result.contentType, 'image/png');
     });
@@ -46,7 +46,7 @@ describe('quality', () => {
 describe('format', () => {
   formats.forEach((value) => {
     it(`should produce an image with format ${value}`, async () => {
-      subject = new Processor(3, `${base}/full/max/0/default.${value}`, streamResolver);
+      subject = new Processor(`${base}/full/max/0/default.${value}`, streamResolver);
       const result = await subject.execute();
       assert.match(result.contentType, /^image\//);
     });
@@ -56,14 +56,14 @@ describe('format', () => {
 describe('region', () => {
   regions.forEach((value) => {
     it(`should produce an image with region ${value}`, async () => {
-      subject = new Processor(3, `${base}/${value}/max/0/default.png`, streamResolver);
+      subject = new Processor(`${base}/${value}/max/0/default.png`, streamResolver);
       const result = await subject.execute();
       assert.strictEqual(result.contentType, 'image/png');
     });
   });
 
   it('should require valid region size', async () => {
-    subject = new Processor(3, `${base}/0,0,0,0/max/0/default.png`, streamResolver);
+    subject = new Processor(`${base}/0,0,0,0/max/0/default.png`, streamResolver);
     assert.rejects(() => subject.execute(), IIIFError);
   });
 });
@@ -71,20 +71,20 @@ describe('region', () => {
 describe('size', () => {
   sizes.forEach((value) => {
     it(`should produce an image with size ${value}`, async () => {
-      subject = new Processor(3, `${base}/full/${value}/0/default.png`, streamResolver);
+      subject = new Processor(`${base}/full/${value}/0/default.png`, streamResolver);
       const result = await subject.execute();
       assert.strictEqual(result.contentType, 'image/png');
     });
   });
 
   it('should require valid size', async () => {
-    subject = new Processor(3, `${base}/full/pct:0/0/default.png`, streamResolver);
+    subject = new Processor(`${base}/full/pct:0/0/default.png`, streamResolver);
     assert.rejects(() => subject.execute(), IIIFError);
   });
 
   it('should select the correct page for the size', async () => {
     let pipeline;
-    subject = new Processor(3, `${base}/full/pct:40/0/default.png`, streamResolver);
+    subject = new Processor(`${base}/full/pct:40/0/default.png`, streamResolver);
     pipeline = await subject.operations(await subject.dimensions()).pipeline();
     assert.strictEqual(pipeline.options.input.page, 1);
   });
@@ -93,7 +93,7 @@ describe('size', () => {
 describe('rotation', () => {
   rotations.forEach((value) => {
     it(`should produce an image with rotation ${value}`, async () => {
-      subject = new Processor(3, `${base}/full/max/${value}/default.png`, streamResolver);
+      subject = new Processor(`${base}/full/max/${value}/default.png`, streamResolver);
       const result = await subject.execute();
       assert.strictEqual(result.contentType, 'image/png');
     });
@@ -103,7 +103,7 @@ describe('rotation', () => {
 describe('IIIF transformation', () => {
   beforeEach(() => {
     consoleWarnMock = jest.spyOn(global.console, 'warn').mockImplementation();
-    subject = new Processor(3, 
+    subject = new Processor(
       `${base}/10,20,30,40/pct:50/45/default.png`,
       streamResolver,
       { dimensionFunction: () => null }
@@ -126,7 +126,7 @@ describe('IIIF transformation', () => {
   
 describe('Two-argument streamResolver', () => {
   beforeEach(() => {
-    subject = new Processor(3, 
+    subject = new Processor(
       `${base}/10,20,30,40/pct:50/45/default.png`,
       ({id, baseUrl}, callback) => { 
         const stream = streamResolver({id, baseUrl});